+++ Example/StripeTerminal.xcframework/ios-arm64_i386_x86_64-simulator/StripeTerminal.framework/Headers/SCPBluetoothConnectionConfiguration.h
@@ -11,6 +11,8 @@
 
 #import <Foundation/Foundation.h>
 
+#import <StripeTerminal/SCPReconnectionDelegate.h>
+
 #import <StripeTerminal/SCPConnectionConfiguration.h>
 
 NS_ASSUME_NONNULL_BEGIN
@@ -39,9 +41,30 @@ NS_SWIFT_NAME(BluetoothConnectionConfiguration)
 @property (nonatomic, readonly) NSString *locationId;
 
 /**
+ When set to true, the Terminal SDK  will attempt a Bluetooth auto-reconnection on any unexpected disconnect.
+ You must also set an `autoReconnectionDelegate` for your application to respond accordingly to reconnection attempts.
+ Enabling `autoReconnectOnUnexpectedDisconnect` without providing an `autoReconnectionDelegate` will error with `SCPErrorReaderConnectionConfigurationInvalid`.
+
+ When set to false, we will immediately surface any disconnection through TerminalDelegate.
+
+ Defaults to false.
+ */
+@property (nonatomic, assign, readonly) BOOL autoReconnectOnUnexpectedDisconnect;
+
+/**
+ Contains callback methods for Bluetooth auto-reconnection.
+ */
+@property (nonatomic, weak, readonly, nullable) id<SCPReconnectionDelegate> autoReconnectionDelegate;
+
+/**
  Initialize your connect options with a location ID.
  */
 - (instancetype)initWithLocationId:(NSString *)locationId;
+
+/**
+ Initialize your connect options with a location ID and an expected behavior for unexpected disconnects.
+ */
+- (instancetype)initWithLocationId:(NSString *)locationId autoReconnectOnUnexpectedDisconnect:(BOOL)autoReconnectOnUnexpectedDisconnect autoReconnectionDelegate:(nullable id<SCPReconnectionDelegate>)autoReconnectionDelegate;
 
 @end